requirements.txt
streamlit
torch
transformers
pillow
numpy
matplotlib
scikit-image
lime
tqdm